+++ packages/core/src/index.ts
@@ -10,6 +10,7 @@ export * from "./plugin.js";
 export * from "./hostInputs.js";
 export * from "./pluginLoader.js";
 export * from "./projectRegistry.js";
+export * from "./repositorySet.js";
 export * from "./runner.js";
 export * from "./runtimeBackends.js";
 export * from "./types.js";